mux-server: extract whatsapp queue retry helpers
